functions/requirements.txt
functions-framework>=3.8.0
firebase-admin>=6.5.0
openai>=1.54.0
firebase-functions
python-dotenv